+++ kovshikov.nikita/S4/treeProcess.hpp
@@ -5,14 +5,14 @@
 
 namespace kovshikov
 {
-  void readData(std::istream& in, Tree< std::string, Tree< long long, std::string > >& allTree);
-  void createTree(Tree< std::string, Tree< long long, std::string > >& allTree, std::string dictionary);
+  void readData(std::istream& in, Tree< std::string, Tree< size_t, std::string > >& allTree);
+  void createTree(Tree< std::string, Tree< size_t, std::string > >& allTree, std::string dictionary);
   bool isDigit(std::string str);
 
-  void print(Tree< std::string, Tree< long long, std::string > >& allTree);
-  void getComplement(Tree< std::string, Tree< long long, std::string > >& allTree);
-  void getIntersect(Tree< std::string, Tree< long long, std::string > >& allTree);
-  void getUnion(Tree< std::string, Tree< long long, std::string > >& allTree);
+  void print(Tree< std::string, Tree< size_t, std::string > >& allTree);
+  void getComplement(Tree< std::string, Tree< size_t, std::string > >& allTree);
+  void getIntersect(Tree< std::string, Tree< size_t, std::string > >& allTree);
+  void getUnion(Tree< std::string, Tree< size_t, std::string > >& allTree);
 
   void outInvalid();
 }
